pyproject.toml
[tool.poetry]
name = "lunchlog"
version = "0.1.0"
description = "Office lunch receipt management and food recommendation system"
authors = ["Higor Rossato <higorleandrorossato@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.2"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.9"
pre-commit = "^3.7.1"
ruff = "^0.5.0"
django-extensions = "^3.2.3"
django-environ = "^0.11.2"
django-storages = "^1.14.3"
boto3 = "^1.34.136"
pillow = "^10.4.0"
requests = "^2.32.3"
drf-spectacular = "^0.27.2"
drf-spectacular-sidecar = "^2024.7.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-django = "^4.8.0"
pytest-cov = "^5.0.0"
black = {version = "^24.4.2", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
